@@ -11,9 +11,8 @@
 	<url>http://shanoir.org</url>
 
 	<properties>
-		<maven.compiler.source>1.8</maven.compiler.source>
-		<maven.compiler.target>1.8</maven.compiler.target>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+		<java.version>11</java.version>
 		<keycloak.version>13.0.1</keycloak.version>
 	</properties>
 
@@ -40,7 +39,7 @@
 		<dependency>
 			<groupId>org.shanoir-ng</groupId>
 			<artifactId>shanoir-ng-exchange</artifactId>
-			<version>0.0.1</version>
+			<version>2.4.5</version>
 			<!-- really important: solves slf4j binding issue and avoid GB of logs during upload with httpClient!!! -->
 			<exclusions>
 			    <exclusion>
@@ -69,7 +68,7 @@
 		<dependency>
 			<groupId>org.shanoir.anonymization</groupId>
 			<artifactId>anonymization</artifactId>
-			<version>0.0.1-SNAPSHOT</version>
+			<version>2.4.5</version>
 			<exclusions>
 				<exclusion>
 					<groupId>org.slf4j</groupId>
@@ -211,19 +210,31 @@
 
 	<repositories>
 		<repository>
-			<id>dcm4che</id>
-			<name>dcm4che Repository</name>
-			<url>http://maven.dcm4che.org</url>
-		</repository>
-		<repository>
 			<id>jboss</id>
 			<name>JBoss Release Repository</name>
-			<url>http://repository.jboss.org/maven2</url>
+			<url>https://repository.jboss.org/maven2</url>
+		</repository>
+		<repository>
+			<id>dcm4che</id>
+			<name>dcm4che</name>
+			<url>https://www.dcm4che.org/maven2/</url>
+			<snapshots>
+				<enabled>true</enabled>
+			</snapshots>
 		</repository>
 	</repositories>
 
 	<build>
 		<plugins>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-compiler-plugin</artifactId>
+				<version>3.8.1</version>
+				<configuration>
+					<source>${java.version}</source>
+					<target>${java.version}</target>
+				</configuration>
+			</plugin>
 			<!-- Place any Ant task here. You can add anything you can add between 
 				<target> and </target> in a build.xml. -->
 			<plugin>
