@@ -2,15 +2,17 @@ package org.shanoir.uploader.action.init;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
-import java.io.File;
 
+import org.apache.log4j.Logger;
+import org.json.JSONException;
 import org.shanoir.uploader.ShUpConfig;
 import org.shanoir.uploader.ShUpOnloadConfig;
 import org.shanoir.uploader.gui.LoginConfigurationPanel;
-import org.shanoir.uploader.service.rest.ShanoirUploaderServiceClientNG;
-import org.shanoir.uploader.service.soap.ServiceConfiguration;
+import org.shanoir.uploader.service.rest.ShanoirUploaderServiceClient;
 
 public class LoginPanelActionListener implements ActionListener {
+	
+	private static Logger logger = Logger.getLogger(LoginPanelActionListener.class);
 
 	private LoginConfigurationPanel loginPanel;
 
@@ -24,9 +26,10 @@ public class LoginPanelActionListener implements ActionListener {
 	public void actionPerformed(ActionEvent e) {
 		String username = this.loginPanel.loginText.getText();
 		String password = String.valueOf(this.loginPanel.passwordText.getPassword());
-		if (ShUpOnloadConfig.isShanoirNg()) {
-			ShanoirUploaderServiceClientNG shanoirUploaderServiceClientNG = ShUpOnloadConfig.getShanoirUploaderServiceClientNG();
-			String token = shanoirUploaderServiceClientNG.loginWithKeycloakForToken(username, password);
+		ShanoirUploaderServiceClient shanoirUploaderServiceClient = ShUpOnloadConfig.getShanoirUploaderServiceClient();
+		String token;
+		try {
+			token = shanoirUploaderServiceClient.loginWithKeycloakForToken(username, password);
 			if (token != null) {
 				ShUpOnloadConfig.setTokenString(token);
 				sSC.getShUpStartupDialog().updateStartupText(
@@ -37,17 +40,11 @@ public class LoginPanelActionListener implements ActionListener {
 						"\n" + ShUpConfig.resourceBundle.getString("shanoir.uploader.startup.test.connection.fail"));
 				sSC.setState(new AuthenticationManualConfigurationState());
 			}
-		} else {
-			ServiceConfiguration serviceConfiguration = ServiceConfiguration.getInstance();
-			serviceConfiguration.setUsername(username);
-			serviceConfiguration.setPassword(password);
-			ShUpConfig.profileProperties.setProperty("shanoir.server.user.name",
-					serviceConfiguration.getUsername());
-			ShUpConfig.profileProperties.setProperty("shanoir.server.user.password",
-					serviceConfiguration.getPassword());
-			final File propertiesFile = new File(ShUpConfig.profileDirectory, ShUpConfig.PROFILE_PROPERTIES);
-			ShUpConfig.encryption.decryptIfEncryptedString(propertiesFile,
-					ShUpConfig.profileProperties, "shanoir.server.user.password");			
+		} catch (JSONException e1) {
+			logger.error(e1.getMessage(), e1);
+			sSC.getShUpStartupDialog().updateStartupText(
+					"\n" + ShUpConfig.resourceBundle.getString("shanoir.uploader.startup.test.connection.fail"));
+			sSC.setState(new AuthenticationManualConfigurationState());
 		}
 		sSC.nextState();
 	}
