@@ -50,7 +50,6 @@ import org.shanoir.uploader.ShUpOnloadConfig;
 import org.shanoir.uploader.action.DownloadOrCopyActionListener;
 import org.shanoir.uploader.action.FindDicomActionListener;
 import org.shanoir.uploader.action.ImportDialogOpener;
-import org.shanoir.uploader.action.ImportDialogOpenerNG;
 import org.shanoir.uploader.action.NoOrYesAnonRChangeListener;
 import org.shanoir.uploader.action.RSDocumentListener;
 import org.shanoir.uploader.action.SelectionActionListener;
@@ -58,7 +57,6 @@ import org.shanoir.uploader.dicom.IDicomServerClient;
 import org.shanoir.uploader.dicom.anonymize.Pseudonymizer;
 import org.shanoir.uploader.exception.PseudonymusException;
 import org.shanoir.uploader.service.rest.UrlConfig;
-import org.shanoir.uploader.service.soap.ServiceConfiguration;
 
 
 /**
@@ -136,7 +134,7 @@ public class MainWindow extends JFrame {
 	public ShUpConfig shanoirUploaderConfiguration;
 	
 	private ImportDialogOpener importDialogOpener;
-	private ImportDialogOpenerNG importDialogOpenerNG;
+	private ImportDialogOpener importDialogOpenerNG;
 
 	/**
 	 * Create the frame.
@@ -204,7 +202,7 @@ public class MainWindow extends JFrame {
 		mnImportExcell.addActionListener(new ActionListener() {
 			@Override
 			public void actionPerformed(ActionEvent e) {
-				ImportFromCSVWindow importcsv = new ImportFromCSVWindow(shanoirUploaderFolder, resourceBundle, scrollPaneUpload, dicomServerClient, ShUpOnloadConfig.getShanoirUploaderServiceClientNG());
+				ImportFromCSVWindow importcsv = new ImportFromCSVWindow(shanoirUploaderFolder, resourceBundle, scrollPaneUpload, dicomServerClient, ShUpOnloadConfig.getShanoirUploaderServiceClient());
 			}
 		});
 
@@ -223,18 +221,6 @@ public class MainWindow extends JFrame {
 				dscw.hostNameLocalPACSTF.setText(ShUpConfig.dicomServerProperties.getProperty("local.dicom.server.host"));
 				dscw.portLocalPACSTF.setText(ShUpConfig.dicomServerProperties.getProperty("local.dicom.server.port"));
 				dscw.aetLocalPACSTF.setText(ShUpConfig.dicomServerProperties.getProperty("local.dicom.server.aet.calling"));
-			}
-		});
-
-		JMenuItem mntmShanoirServerConfiguration = new JMenuItem(
-				resourceBundle.getString("shanoir.uploader.configurationMenu.shanoirServer"));
-		mnConfiguration.add(mntmShanoirServerConfiguration);
-
-		mntmShanoirServerConfiguration.addActionListener(new ActionListener() {
-			@Override
-			public void actionPerformed(ActionEvent arg0) {
-				ShanoirServerConfigurationWindow sscw = new ShanoirServerConfigurationWindow(shanoirUploaderFolder,
-						ServiceConfiguration.getInstance(), resourceBundle);
 			}
 		});
 
@@ -810,7 +796,6 @@ public class MainWindow extends JFrame {
 		 */
 
 		importDialogOpener = new ImportDialogOpener(this, ShUpOnloadConfig.getShanoirUploaderServiceClient());
-		importDialogOpenerNG = new ImportDialogOpenerNG(this, ShUpOnloadConfig.getShanoirUploaderServiceClientNG());
 		
 		// add ShUp principal panel (splitPane) and upload job display pane
 		// (scrollPaneUpload) to TabbedPane
@@ -927,11 +912,11 @@ public class MainWindow extends JFrame {
 		this.importDialogOpener = importDialogOpener;
 	}
 
-	public ImportDialogOpenerNG getImportDialogOpenerNG() {
+	public ImportDialogOpener getImportDialogOpenerNG() {
 		return importDialogOpenerNG;
 	}
 
-	public void setImportDialogOpenerNG(ImportDialogOpenerNG importDialogOpenerNG) {
+	public void setImportDialogOpenerNG(ImportDialogOpener importDialogOpenerNG) {
 		this.importDialogOpenerNG = importDialogOpenerNG;
 	}
 	
